move resolving of tags
