+++ client/src/lib/components/Cards/Cards.library.tsx
@@ -19,7 +19,7 @@ export const AllCards = () => (
     <Card
       title="A nice title"
       duration={40}
-      tagIds={['43a6dc0d-d866-4b79-9453-13c44e906128']}
+      tags={[{tag: 'Gratitude', featuredOrder: 1, id: 'some-id'}]}
       hostPictureURL="https://res.cloudinary.com/twentyninek/image/upload/q_auto,t_global/v1636016815/Singles/sticky_eng_ps00eg.png"
       hostName="Jenny"
       image={{
@@ -33,6 +33,7 @@ export const AllCards = () => (
     <Spacer16 />
     <Card
       title="A nice title"
+      tags={[]}
       duration={40}
       hostPictureURL="https://res.cloudinary.com/twentyninek/image/upload/q_auto,t_global/v1636016815/Singles/sticky_eng_ps00eg.png"
       hostName="Jenny WithALongName"
@@ -48,6 +49,7 @@ export const AllCards = () => (
     <Card
       title="A longer title and animation"
       lottie={lottieSource}
+      tags={[]}
       image={{
         uri: 'https://res.cloudinary.com/twentyninek/image/upload/v1646061249/Illustrations_Tests/take-test_c4qa3u.png',
       }}
@@ -59,6 +61,7 @@ export const AllCards = () => (
 
     <Card
       title="With reminder and menu"
+      tags={[]}
       image={{
         uri: 'https://res.cloudinary.com/twentyninek/image/upload/q_auto,t_global/v1636016815/Singles/sticky_eng_ps00eg.png',
       }}
@@ -69,6 +72,7 @@ export const AllCards = () => (
     <Spacer16 />
     <Card
       title="Private session starts soon"
+      tags={[]}
       image={{
         uri: 'https://res.cloudinary.com/twentyninek/image/upload/q_auto,t_global/v1636016815/Singles/sticky_eng_ps00eg.png',
       }}
@@ -83,6 +87,7 @@ export const AllCards = () => (
     <Spacer16 />
     <Card
       title="Session starts in 24h"
+      tags={[]}
       image={{
         uri: 'https://res.cloudinary.com/twentyninek/image/upload/q_auto,t_global/v1636016815/Singles/sticky_eng_ps00eg.png',
       }}
